Recipes_app/recipes_app/requirements.txt
Django==4.2.3
psycopg2==2.9.6
flake8==6.1.0
Pillow==10.0.0
djangorestframework==3.14.0